+++ CHANGELOG.md
@@ -50,3 +50,7 @@ New themes
 # 1.2.3
 Fix color of the edit text when unfocus.
 New component! Long edit text.
+
+# 1.2.40
+Long text will shrink and extend upon the text's length,
+Fix description space.